Show agent status indicator in all worktree panes (#62)
Extract shared worktree_list_item() into common.rs so the dashboard and
repo-detail worktree panes render identically. Previously each view
duplicated the status badge, ticket info, and agent indicator logic.
