+++ conductor-tui/src/ui/common.rs
@@ -1,7 +1,8 @@
+use conductor_core::worktree::Worktree;
 use ratatui::layout::Rect;
 use ratatui::style::{Color, Modifier, Style};
 use ratatui::text::{Line, Span};
-use ratatui::widgets::Paragraph;
+use ratatui::widgets::{ListItem, Paragraph};
 use ratatui::Frame;
 
 use crate::state::{AppState, View};
@@ -83,6 +84,92 @@ pub fn render_status_bar(frame: &mut Frame, area: Rect, state: &AppState) {
         Style::default().fg(Color::DarkGray),
     )));
     frame.render_widget(bar, area);
+}
+
+/// Build a `ListItem` for a worktree row.
+///
+/// Both the dashboard and repo-detail worktree panes use this so the
+/// format stays consistent.  Pass `repo_prefix` to prepend the repo
+/// slug (dashboard style) and `show_branch` to append the branch name
+/// (repo-detail style).
+pub fn worktree_list_item(
+    wt: &Worktree,
+    state: &AppState,
+    repo_prefix: Option<&str>,
+    show_branch: bool,
+) -> ListItem<'static> {
+    let is_active = wt.is_active();
+    let status_color = match wt.status.as_str() {
+        "active" => Color::Green,
+        "merged" => Color::Blue,
+        "abandoned" => Color::Red,
+        _ => Color::White,
+    };
+    let text_style = if is_active {
+        Style::default()
+    } else {
+        Style::default().fg(Color::DarkGray)
+    };
+
+    let mut spans: Vec<Span<'static>> = Vec::new();
+
+    if let Some(prefix) = repo_prefix {
+        spans.push(Span::styled(
+            format!("{prefix}/"),
+            Style::default().fg(Color::DarkGray),
+        ));
+    }
+
+    spans.push(Span::styled(
+        wt.slug.clone(),
+        text_style.add_modifier(if is_active {
+            Modifier::BOLD
+        } else {
+            Modifier::DIM
+        }),
+    ));
+
+    if show_branch {
+        spans.push(Span::styled(format!("  {}", wt.branch), text_style));
+    }
+
+    spans.push(Span::raw("  "));
+    spans.push(Span::styled(
+        format!("[{}]", wt.status),
+        Style::default().fg(status_color),
+    ));
+
+    if let Some(ticket) = wt
+        .ticket_id
+        .as_ref()
+        .and_then(|tid| state.data.ticket_map.get(tid))
+    {
+        let ticket_state_color = match ticket.state.as_str() {
+            "open" => Color::Green,
+            "closed" => Color::DarkGray,
+            "in_progress" => Color::Yellow,
+            _ => Color::White,
+        };
+        spans.push(Span::raw("  "));
+        spans.push(Span::styled(
+            format!("#{} {}", ticket.source_id, ticket.state),
+            Style::default().fg(ticket_state_color),
+        ));
+    }
+
+    if let Some(run) = state.data.latest_agent_runs.get(&wt.id) {
+        let (symbol, color) = match run.status.as_str() {
+            "running" => ("● running", Color::Yellow),
+            "completed" => ("✓ completed", Color::Green),
+            "failed" => ("✗ failed", Color::Red),
+            "cancelled" => ("○ cancelled", Color::DarkGray),
+            _ => ("? unknown", Color::White),
+        };
+        spans.push(Span::raw("  "));
+        spans.push(Span::styled(symbol, Style::default().fg(color)));
+    }
+
+    ListItem::new(Line::from(spans))
 }
 
 /// Calculate elapsed time from an ISO 8601 timestamp to now.
